requirements.txt
asgiref==3.8.1
Django==5.1.3
django-ckeditor-5==0.2.15
django-environ==0.11.2
django-js-asset==2.2.0
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.2
tzdata==2024.2
